Move SendFlows to separate struct for more obvious usage
