+++ send.go
@@ -20,16 +20,11 @@ import (
 	"github.com/kentik/libkflow/metrics"
 )
 
-// messagePrefix is an 80-byte prefix for the message header when sending kflow to the Kentik API. This is a deprecated
-// header, but the bytes must remain for backwards compatibility with the Kentik API.
-var messagePrefix = [80]byte{}
-
 // A Sender aggregates and transmits flow information to Kentik.
 type Sender struct {
 	agg               *agg.Agg
 	exit              chan struct{}
 	url               *url.URL
-	timeout           time.Duration
 	client            *api.Client
 	sample            int
 	ticker            *time.Ticker
@@ -43,12 +38,11 @@ type Sender struct {
 	Metrics           *metrics.Metrics
 }
 
-func newSender(url *url.URL, timeout time.Duration) *Sender {
+func newSender(url *url.URL) *Sender {
 	tickerCtx, cancelFunc := context.WithCancel(context.Background())
 	return &Sender{
 		exit:             make(chan struct{}),
 		url:              url,
-		timeout:          timeout,
 		ticker:           time.NewTicker(20 * time.Minute),
 		tickerCtx:        tickerCtx,
 		tickerCancelFunc: cancelFunc,
@@ -60,80 +54,6 @@ func (s *Sender) Send(flow *flow.Flow) {
 	log.Debugf("sending flow to aggregator")
 	flow.DeviceId = uint32(s.Device.ID)
 	s.agg.Add(flow)
-}
-
-// SendFlows sends the flows to the Kentik API, returning the number of bytes sent as the payload. The device ID on
-// the flows is set to the device ID of the sender, regardless of what it was previously set to. This is to ensure all
-// data matches the expectations of the downstream URL/API.
-//
-// This will directly send the slice of flows without any additional downsampling or rate limiting. This does not
-// contribute to the underlying Send call.
-func (s *Sender) SendFlows(flows []flow.Flow) (int64, error) {
-	s.workers.Add(1)
-	defer s.workers.Done()
-
-	if s.Device == nil {
-		return 0, fmt.Errorf("device not initialized")
-	}
-	if len(flows) == 0 {
-		return 0, nil
-	}
-	decoratedURL, err := s.createURLString()
-	if err != nil {
-		return 0, fmt.Errorf("failed to create URL string: %w", err)
-	}
-
-	if s.Metrics != nil {
-		s.Metrics.TotalFlowsIn.Mark(int64(len(flows)))
-	}
-
-	// ensure all flows have the device ID set; otherwise it may not be properly queried
-	for i := range flows {
-		flows[i].DeviceId = uint32(s.Device.ID)
-	}
-
-	// ensure the sample rate is matching the kentik api expectations
-	flow.NormalizeSampleRate(flows, 0)
-
-	// serialize the data
-	_, segment, err := capnp.NewMessage(capnp.SingleSegment(nil))
-	if err != nil {
-		return 0, fmt.Errorf("failed to create capn proto segment: %w", err)
-	}
-	message, err := flow.ToCapnProtoMessage(flows, segment)
-	if err != nil {
-		return 0, fmt.Errorf("failed to convert flows to capn proto: %w", err)
-	}
-
-	// write the data with additional gzip compression
-	buf := &bytes.Buffer{}
-	z := gzip.NewWriter(buf)
-	_, err = z.Write(messagePrefix[:])
-	if err != nil {
-		return 0, fmt.Errorf("failed to write empty message header: %w", err)
-	}
-	err = capnp.NewPackedEncoder(z).Encode(message)
-	if err != nil {
-		return 0, fmt.Errorf("failed to encode packed capn proto message: %w", err)
-	}
-	err = z.Close()
-	if err != nil {
-		return 0, fmt.Errorf("failed to close gzip writer: %w", err)
-	}
-
-	// send the compressed and packed message to the Kentik API
-	payloadLength := int64(len(buf.Bytes()))
-	err = s.client.SendFlow(decoratedURL, buf)
-	if err != nil {
-		return 0, err
-	}
-
-	if s.Metrics != nil {
-		s.Metrics.TotalFlowsOut.Mark(int64(len(flows)))
-		s.Metrics.BytesSent.Mark(payloadLength)
-	}
-
-	return payloadLength, nil
 }
 
 // Stop requests a graceful shutdown of the Sender.
@@ -336,12 +256,5 @@ func (s *Sender) createURLString() (string, error) {
 	if s.url == nil {
 		return "", fmt.Errorf("url not initialized")
 	}
-
-	// Create a new URL to avoid modifying the original, backed by the config
-	u := *s.url
-	q := u.Query()
-	q.Set("sid", "0")
-	q.Set("sender_id", s.Device.ClientID())
-	u.RawQuery = q.Encode()
-	return u.String(), nil
+	return createURLString(*s.url, s.Device.ClientID()), nil
 }
