+++ send_test.go
@@ -42,38 +42,6 @@ func TestSender(t *testing.T) {
 	assert.Equal(flowToCHF(expected, t).String(), msgs.At(0).String())
 }
 
-func TestSender_SendFlows(t *testing.T) {
-	sender, server, assert := setup(t)
-
-	expected1 := flow.Flow{
-		DeviceId:  uint32(sender.Device.ID),
-		SrcAs:     rand.Uint32(),
-		DstAs:     rand.Uint32(),
-		SampleAdj: true,
-	}
-	expected2 := flow.Flow{
-		DeviceId:  uint32(sender.Device.ID),
-		SrcAs:     rand.Uint32(),
-		DstAs:     rand.Uint32(),
-		SampleAdj: true,
-	}
-
-	flows := []flow.Flow{expected1, expected2}
-	n, err := sender.SendFlows(flows)
-	assert.NoError(err)
-
-	msgs, err := receive(server)
-	if err != nil {
-		t.Fatal(err)
-	}
-	assert.Equal(int64(len(flows)), sender.Metrics.TotalFlowsIn.Count())
-	assert.Equal(int64(len(flows)), sender.Metrics.TotalFlowsOut.Count())
-	assert.Equal(n, sender.Metrics.BytesSent.Count())
-	assert.Equal(len(flows), msgs.Len())
-	assert.Equal(flowToCHF(expected1, t).String(), msgs.At(0).String())
-	assert.Equal(flowToCHF(expected2, t).String(), msgs.At(1).String())
-}
-
 func TestSenderStop(t *testing.T) {
 	sender, _, assert := setup(t)
 	stopped := sender.Stop(100 * time.Millisecond)
@@ -140,7 +108,7 @@ func setup(t testing.TB) (*Sender, *test.Server, *assert.Assertions) {
 	server.Log.SetOutput(io.Discard)
 
 	url := server.URL(test.FLOW)
-	sender := newSender(url, 1*time.Second)
+	sender := newSender(url)
 	sender.Metrics = metrics
 	sender.start(agg, client, device, 1)
 
@@ -226,7 +194,7 @@ func TestSender_createURLString(t *testing.T) {
 	agg, err := agg.NewAgg(10*time.Millisecond, 100, m)
 	require.NoError(t, err)
 
-	sender := newSender(u, 1*time.Second)
+	sender := newSender(u)
 	err = sender.start(agg, client, device, 1)
 	require.NoError(t, err)
